+++ core/vm/arch/memory.h
@@ -230,7 +230,7 @@ class MemoryManager {
   static size_t* AllocateArray(const long size, const MemoryType type, size_t* op_stack, long stack_pos, bool collect = true);
   
   // object verification
-  static size_t* ValidObjectCast(size_t* mem, long to_id, int* cls_hierarchy, int** cls_interfaces);
+  static size_t* ValidObjectCast(size_t* mem, long to_id, long* cls_hierarchy, long** cls_interfaces);
   
   //
   // returns the class reference for an object instance
